+++ CHANGELOG.md
@@ -3,6 +3,109 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.105](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.104...v2.0.0-beta.105) (2020-10-16)
+
+
+### Features
+
+* **alert-base:** create AlertBase component to share internally ([e2e2dfc](http://stash.cengage.com:7999/FRONT/react-magma/commits/e2e2dfcc9ace426aea82cc6102affe30e5500796))
+
+
+### BREAKING CHANGES
+
+* **alert-base:** removed undocumented props used only internally from Alert
+
+
+
+
+
+# [2.0.0-beta.104](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.103...v2.0.0-beta.104) (2020-10-16)
+
+
+### Features
+
+* **base-input:** BaseInput component renamed to InputBase ([8309338](http://stash.cengage.com:7999/FRONT/react-magma/commits/8309338873875d9fcc7baa537db0c44a49d37bcf))
+* **base-input:** move InputBase export for alphabetical context ([51651c0](http://stash.cengage.com:7999/FRONT/react-magma/commits/51651c071d3269c0d9fb7966cb42ad61c5cee6f2))
+
+
+### BREAKING CHANGES
+
+* **base-input:** BaseInput component name changed
+
+
+
+
+
+# [2.0.0-beta.103](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.102...v2.0.0-beta.103) (2020-10-16)
+
+
+### Features
+
+* **hyperlink:** Change name from HyperLink to Hyperlink ([5aba7dc](http://stash.cengage.com:7999/FRONT/react-magma/commits/5aba7dcef9c4af060bd373f46e45c94ad45eb6c9))
+
+
+### BREAKING CHANGES
+
+* **hyperlink:** component name change for Hyperlink
+
+
+
+
+
+# [2.0.0-beta.102](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.101...v2.0.0-beta.102) (2020-10-16)
+
+
+### Bug Fixes
+
+* **visuallyhidden:** add testid, also add testid to missing prop tables ([917bf1c](http://stash.cengage.com:7999/FRONT/react-magma/commits/917bf1c8feedc93c7c445016363054a45c213c5e))
+
+
+
+
+
+# [2.0.0-beta.101](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.100...v2.0.0-beta.101) (2020-10-16)
+
+**Note:** Version bump only for package react-magma
+
+
+
+
+
+# [2.0.0-beta.100](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.99...v2.0.0-beta.100) (2020-10-16)
+
+
+### Features
+
+* **multi:** change isDisabled prop to disabled ([9d5c57b](http://stash.cengage.com:7999/FRONT/react-magma/commits/9d5c57bd802f9dcf7cac58deda60dd8843087457))
+* **radio-group:** isRequired prop renamed to required ([aae29e3](http://stash.cengage.com:7999/FRONT/react-magma/commits/aae29e3ffb6fb6c5235473d0d10d75b0f7157fac))
+
+
+### BREAKING CHANGES
+
+* **radio-group:** isRequired prop changed to required in RadioGroup
+* **multi:** prop name change
+
+
+
+
+
+# [2.0.0-beta.99](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.98...v2.0.0-beta.99) (2020-10-16)
+
+
+### Bug Fixes
+
+* **docs:** change docs main-nav to buttons ([544c9d3](http://stash.cengage.com:7999/FRONT/react-magma/commits/544c9d37575fc800887ea42c16a3c18d64b33af3))
+* **docs:** fix skiplinks in doc to not prevent any clicks ([af2bd7b](http://stash.cengage.com:7999/FRONT/react-magma/commits/af2bd7b4d92c8c21f89bf6e95edce73e03849fbb))
+* **docs:** fix subnavs for doc sections ([f685a9e](http://stash.cengage.com:7999/FRONT/react-magma/commits/f685a9eb853e5f06451898991c7385f41bb65c76))
+* **docs:** maintain sidebar location ([665362d](http://stash.cengage.com:7999/FRONT/react-magma/commits/665362da26d2226202e7bf960d3e6692c4d54c26))
+* **docs:** move skiplink back up to first ([cd68797](http://stash.cengage.com:7999/FRONT/react-magma/commits/cd687970be84d82244669fdb8591d60bf33fb65b))
+* **docs:** set focus and scroll to active section when clicking nav bar ([60d5f9a](http://stash.cengage.com:7999/FRONT/react-magma/commits/60d5f9a04ec1ab0046bd7efa84406ced8956c264))
+* **docs:** switch to event handler for focus and movement ([4d15af8](http://stash.cengage.com:7999/FRONT/react-magma/commits/4d15af85a43f613b39719e7b3e6cb7539fc21d1c))
+
+
+
+
+
 # [2.0.0-beta.98](http://stash.cengage.com:7999/FRONT/react-magma/compare/v2.0.0-beta.97...v2.0.0-beta.98) (2020-10-14)